arithmetic expansion in assignment
