+++ Makefile
@@ -231,6 +231,7 @@ HISTORY_MANAGEMENT := $(addprefix $(HISTORY_MANAGEMENT_DIR)/,$(HISTORY_MANAGEMEN
 # Arithmetic Expansion
 ARITHMETIC_EXPANSION_DIR := $(FEATURES_DIR)/arithmetic_expansion
 ARITHMETIC_EXPANSION_FILE += arx.c
+ARITHMETIC_EXPANSION_FILE += arx_handler.c
 ARITHMETIC_EXPANSION_FILE += arx_check.c
 ARITHMETIC_EXPANSION_FILE += arx_exp.c
 ARITHMETIC_EXPANSION_FILE += arx_lexer.c
